added custom hook for getting wallet and setting address to global state
